requirements.txt
# Core MCP dependencies
fastmcp
mcp

# HTTP client for API requests
httpx

# FastAPI and server dependencies
fastapi
uvicorn
pydantic

# Google APIs (Gmail, Calendar)
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# File handling and utilities
pathlib
# mimetypes

# Date/time handling
python-dateutil